requirements.txt
streamlit==1.32.0
pandas==2.2.0
openpyxl==3.1.2
pdfplumber==0.10.3
pytesseract==0.3.10
python-dotenv==1.0.0
plotly==5.18.0
pillow==10.2.0
openai==1.12.0
sqlite3==3.35.0
xero-python==1.25.0
reportlab==4.1.0
numpy==1.26.0
scikit-learn==1.4.0
python-dateutil==2.8.2
schedule==1.2.1 